+++ internal/pkg/util/env/env.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2018, Sylabs Inc. All rights reserved.
+// Copyright (c) 2018-2023, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -6,9 +6,12 @@
 package env
 
 import (
+	"context"
 	"fmt"
 	"os"
 	"strings"
+
+	"github.com/sylabs/singularity/v4/internal/pkg/util/shell/interpreter"
 )
 
 // SetFromList sets environment variables from environ argument list.
@@ -23,4 +26,72 @@ func SetFromList(environ []string) error {
 		}
 	}
 	return nil
+}
+
+// SingularityEnvMap returns a map of SINGULARITYENV_ prefixed env vars to their values.
+func SingularityEnvMap() map[string]string {
+	singularityEnv := map[string]string{}
+
+	for _, envVar := range os.Environ() {
+		if !strings.HasPrefix(envVar, SingularityEnvPrefix) {
+			continue
+		}
+		parts := strings.SplitN(envVar, "=", 2)
+		if len(parts) < 2 {
+			continue
+		}
+		key := strings.TrimPrefix(parts[0], SingularityEnvPrefix)
+		singularityEnv[key] = parts[1]
+	}
+
+	return singularityEnv
+}
+
+// FileMap returns a map of KEY=VAL env vars from an environment file f. The env
+// file is shell evaluated using mvdan/sh with arguments and environment set
+// from args and currentEnv.
+func FileMap(ctx context.Context, f string, args []string, currentEnv []string) (map[string]string, error) {
+	envMap := map[string]string{}
+
+	content, err := os.ReadFile(f)
+	if err != nil {
+		return envMap, fmt.Errorf("could not read environment file %q: %w", f, err)
+	}
+
+	// Use the embedded shell interpreter to evaluate the env file, with an empty starting environment.
+	// Shell takes care of comments, quoting etc. for us and keeps compatibility with native runtime.
+	env, err := interpreter.EvaluateEnv(ctx, content, args, currentEnv)
+	if err != nil {
+		return envMap, fmt.Errorf("while processing %s: %w", f, err)
+	}
+
+	for _, envVar := range env {
+		parts := strings.SplitN(envVar, "=", 2)
+		if len(parts) < 2 {
+			continue
+		}
+		// Strip out the runtime env vars set by the shell interpreter so we
+		// don't attempt to overwrite bash builtin readonly vars.
+		// https://github.com/sylabs/singularity/issues/1263
+		if parts[0] == "GID" ||
+			parts[0] == "HOME" ||
+			parts[0] == "IFS" ||
+			parts[0] == "OPTIND" ||
+			parts[0] == "PWD" ||
+			parts[0] == "UID" {
+			continue
+		}
+		envMap[parts[0]] = parts[1]
+	}
+
+	return envMap, nil
+}
+
+// MergeMap merges two maps of environment variables, with values in b replacing
+// values also set in a.
+func MergeMap(a map[string]string, b map[string]string) map[string]string {
+	for k, v := range b {
+		a[k] = v
+	}
+	return a
 }